[Common-TextInput] Added common classes for TextInput and PickerInput.
