+++ source/reactnative/src/components/inputs/date-input.js
@@ -6,24 +6,16 @@
 
 import React from 'react';
 import {
-    View,
-    StyleSheet,
-    TouchableOpacity,
     DatePickerIOS,
     DatePickerAndroid,
     Platform,
-    TextInput,
-    Modal,
-    Button,
-    ViewPropTypes
 } from 'react-native';
 
-import Input from './input';
 import { InputContextConsumer } from '../../utils/context'
 import * as Constants from '../../utils/constants';
 import { StyleManager } from '../../styles/style-config'
 import { HostConfigManager } from '../../utils/host-config'
-
+import { PickerInput } from './picker-input'
 
 export class DateInput extends React.Component {
 
@@ -33,10 +25,6 @@ export class DateInput extends React.Component {
         super(props);
 
         this.payload = props.json;
-        this.id = Constants.EmptyString;
-        this.placeHolder = Constants.EmptyString;
-        this.type = Constants.EmptyString;
-        this.modalButtonText = Constants.DoneString
         this.parseHostConfig();
     }
 
@@ -44,9 +32,6 @@ export class DateInput extends React.Component {
     * @description Parse hostconfig specific to this element
     */
     parseHostConfig() {
-        this.id = this.payload.id;
-        this.type = this.payload.type;
-        this.placeholder = this.payload.placeholder;
         this.state = {
             chosenDate: this.payload.value ? this.parseDateString(this.payload.value) : new Date(),
             minDate: this.payload.min ? this.parseDateString(this.payload.min) : undefined,
@@ -137,102 +122,25 @@ export class DateInput extends React.Component {
             return null;
         }
 
-        const {
-            id,
-            type,
-            placeholder,
-            modalButtonText
-        } = this;
-
-        if (!id || !type) {
-            return null;
-        }
-
-        modalOverlayStyle = ViewPropTypes.style,
-            modalStyle = ViewPropTypes.style,
-            modalButtonStyle = ViewPropTypes.style,
-            modalBtnContainer = ViewPropTypes.style
-
         return (
             <InputContextConsumer>
                 {({ addInputItem }) => (
-                    <Input json={this.payload}>
-
-                        <TouchableOpacity style={styles.inputWrapper} onPress={this.showDatePicker}>
-                            {/* added extra view to fix touch event in ios . */}
-                            <View pointerEvents='none' >
-                                <TextInput
-                                    style={[styles.input, this.styleConfig.fontConfig]}
-                                    autoCapitalize={Constants.NoneString}
-                                    autoCorrect={false}
-                                    placeholder={placeholder}
-                                    textContentType={Constants.NoneString}
-                                    underlineColorAndroid={Constants.TransparentString}
-                                    value={this.state.value}>
-                                    {addInputItem(this.id, this.state.value)}
-                                </TextInput>
-                            </View>
-                        </TouchableOpacity>
-                        <Modal
-                            animationType='slide'
-                            transparent
-                            visible={this.state.modalVisible}
-                            onRequestClose={this.handleModalClose}>
-                            <View style={[styles.overlay, modalOverlayStyle]}>
-                                <View style={[styles.modal, modalStyle]}>
-                                    <View style={[styles.modalBtnContainer, modalBtnContainer]}>
-                                        <Button
-                                            style={[modalButtonStyle]}
-                                            title={modalButtonText}
-                                            onPress={this.handleModalClose}
-                                        />
-                                    </View>
-                                    <DatePickerIOS
-                                        mode='date'
-                                        date={this.state.chosenDate || new Date()}
-                                        minimumDate={this.state.minDate}
-                                        maximumDate={this.state.maxDate}
-                                        onDateChange={this.handleDateChange} />
-                                </View>
-                            </View>
-                        </Modal>
-                    </Input>)}
+					<PickerInput
+						json={this.payload}
+						value={this.state.value}
+						showPicker={this.showDatePicker}
+						addInputItem={addInputItem} 
+						modalVisible={this.state.modalVisible}
+						handleModalClose={this.handleModalClose}
+						chosenDate={this.state.chosenDate || new Date()}
+						minDate={this.state.minDate}
+						maxDate={this.state.maxDate}
+						handleDateChange={this.handleDateChange} 
+						mode='date'
+					/>
+				)}
             </InputContextConsumer>
         );
     }
 }
 
-const styles = StyleSheet.create({
-    inputWrapper: {
-        width: Constants.FullWidth,
-        marginTop: 15,
-    },
-    input: {
-        width: Constants.FullWidth,
-        height: 44,
-        padding: 5,
-        borderWidth: 1,
-        backgroundColor: Constants.WhiteColor,
-        borderColor: Constants.LightGreyColor,
-        borderRadius: 5,
-    },
-    overlay: {
-        flex: 1,
-        backgroundColor: 'rgba(0,0,0,.3)',
-        alignItems: Constants.CenterString,
-        justifyContent: Constants.FlexEnd
-    },
-    modal: {
-        backgroundColor: Constants.WhiteColor,
-        height: 260,
-        width: Constants.FullWidth
-    },
-    modalBtnContainer: {
-        width: Constants.FullWidth,
-        alignItems: Constants.CenterString,
-        flexDirection: Constants.FlexRow,
-        justifyContent: Constants.FlexEnd,
-        paddingHorizontal: 15,
-        marginTop: 15
-    }
-});
