+++ source/reactnative/src/components/inputs/number-input.js
@@ -5,17 +5,13 @@
  */
 
 import React from "react";
-import {
-    StyleSheet,
-    TextInput
-} from 'react-native';
-
 import { DismissKeyboardView } from '../containers/dismiss-keyboard';
 import { InputContextConsumer } from '../../utils/context'
 import Input from './input';
 import * as Constants from '../../utils/constants';
 import { StyleManager } from "../../styles/style-config";
 import { HostConfigManager } from '../../utils/host-config'
+import { TextInputWrapper } from './text-input-wrapper'
 
 
 const NUM_REGEX = /^[0-9][\.\d]*(,\d+)?$/;
@@ -29,13 +25,6 @@ export class NumberInput extends React.Component {
 
         this.payload = props.json;
         this.id = Constants.EmptyString;
-        this.isMultiline = false;
-        this.maxlength = 0;
-        this.placeHolder = Constants.EmptyString;
-        this.style = Constants.EmptyString;
-        this.type = Constants.EmptyString;
-        this.value = Constants.EmptyString;
-        this.keyboardType = Constants.EmptyString;
         this.state = {
             isError: false,
             numberValue: Constants.EmptyString,
@@ -50,70 +39,31 @@ export class NumberInput extends React.Component {
 
         this.parseHostConfig();
 
-        const {
-            id,
-            type,
-            placeholder,
-            keyboardType
-        } = this;
-
-        if (!id || !type) {
-            return null;
-        }
-
         return (
             <InputContextConsumer>
                 {({ addInputItem }) => (
-                    <Input json={this.payload}>
-                        <DismissKeyboardView>
-                            <TextInput
-                                style={this.getComputedStyles()}
-                                autoCapitalize={Constants.NoneString}
-                                autoCorrect={false}
-                                placeholder={placeholder}
-                                underlineColorAndroid={Constants.TransparentString}
-                                clearButtonMode={Constants.WhileEditingString}
-                                textContentType={Constants.NoneString}
-                                keyboardType={keyboardType}
-                                onFocus={this.handleFocus}
-                                onBlur={this.handleBlur}
-                                onChangeText={(text) => this.onTextChanged(text, addInputItem)}
-                                value={this.state.numberValue}
-                            />
-                        </DismissKeyboardView>
-                    </Input>)}
+                 	<DismissKeyboardView>
+						<TextInputWrapper 
+							json={this.payload}
+							handleFocus={this.handleFocus}
+							handleBlur={this.handleBlur}
+							textValueChanged={ (text) => { this.onTextChanged(text, addInputItem) }}
+							value={this.state.numberValue} 
+							isError={this.state.isError}     
+						/>
+				 	</DismissKeyboardView>
+				)}
             </InputContextConsumer>
 
         );
     }
 
     /**
-     * @description Return the input styles applicable based on the given payload
-     */
-    getComputedStyles = () => {
-        const { isMultiline } = this;
-
-        let inputComputedStyles = [styles.input, this.styleConfig.fontConfig];
-        isMultiline ?
-            inputComputedStyles.push(styles.multiLineHeight) :
-            inputComputedStyles.push(styles.singleLineHeight);
-        this.state.isError ?
-            inputComputedStyles.push(styles.withBorderColorRed) :
-            inputComputedStyles.push(styles.withBorderColor);
-
-        return inputComputedStyles;
-    }
-
-    /**
      * @description Parse hostconfig specific to this element
      */
     parseHostConfig() {
-        this.id = this.payload.id;
-        this.type = this.payload.type;
         this.min = this.payload.min;
         this.max = this.payload.max;
-        this.placeholder = this.payload.placeholder;
-        this.keyboardType = Constants.NumericString;
     }
 
     /**
@@ -167,28 +117,5 @@ export class NumberInput extends React.Component {
         return true
     }
 };
-
-const styles = StyleSheet.create({
-    withBorderColor: {
-        borderColor: Constants.LightGreyColor,
-    },
-    withBorderColorRed: {
-        borderColor: '#cc3300',
-    },
-    multiLineHeight: {
-        height: 88,
-    },
-    singleLineHeight: {
-        height: 44,
-    },
-    input: {
-        width: Constants.FullWidth,
-        padding: 5,
-        marginTop: 15,
-        borderWidth: 1,
-        backgroundColor: Constants.WhiteColor,
-        borderRadius: 5,
-    },
-});
