+++ source/reactnative/src/components/inputs/text-input.js
@@ -5,11 +5,6 @@
  */
 
 import React from "react";
-import {
-    StyleSheet,
-    TextInput
-} from 'react-native';
-
 import Input from './input';
 import { StyleManager } from "../../styles/style-config";
 import * as Utils from '../../utils/util';
@@ -17,6 +12,7 @@ import * as Enums from '../../utils/enums';
 import * as Constants from '../../utils/constants';
 import { InputContextConsumer } from '../../utils/context'
 import { HostConfigManager } from '../../utils/host-config'
+import { TextInputWrapper } from './text-input-wrapper'
 
 const EMAIL_REGEX = /^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/;
 const URL_REGEX = /((([A-Za-z]{3,9}:(?:\/\/)?)(?:[\-;:&=\+\$,\w]+@)?[A-Za-z0-9\.\-]+|(?:www\.|[\-;:&=\+\$,\w]+@)[A-Za-z0-9\.\-]+)((?:\/[\+~%\/\.\w\-_]*)?\??(?:[\-\+=&;%@\.\w_]*)#?(?:[\.\!\/\\\w]*))?)/;
@@ -29,18 +25,11 @@ export class InputText extends React.Component {
     constructor(props) {
         super(props);
 
-        this.payload = props.json;
-        this.id = Constants.EmptyString;
-        this.isMultiline = Boolean;
-        this.maxlength = 0;
-        this.placeHolder = Constants.EmptyString;
-        this.style = Constants.EmptyString;
-        this.type = Constants.EmptyString;
-        this.value = Constants.EmptyString;
-        this.keyboardType = Constants.EmptyString;
+		this.payload = props.json;
+		this.id = props.json.id;
         this.state = {
             isError: false,
-            text: '',
+            text: Constants.EmptyString,
         }
     }
 
@@ -59,63 +48,21 @@ export class InputText extends React.Component {
             return null;
         }
 
-        this.parseHostConfig();
-
-        const {
-            id,
-            type,
-            isMultiline,
-            placeholder,
-            maxLength,
-            style,
-            keyboardType
-        } = this;
-
-        if (!id || !type) {
-            return null;
-        }
-
         return (
             <InputContextConsumer>
                 {({ addInputItem }) => (
-                    <Input json={this.payload}>
-                        <TextInput
-                            style={this.getComputedStyles()}
-                            autoCapitalize={Constants.NoneString}
-                            autoCorrect={false}
-                            placeholder={placeholder}
-                            multiline={isMultiline}
-                            maxLength={maxLength}
-                            underlineColorAndroid={Constants.TransparentString}
-                            clearButtonMode={Constants.WhileEditingString}
-                            textContentType={style}
-                            keyboardType={keyboardType}
-                            onFocus={this.handleFocus}
-                            onBlur={this.handleBlur}
-                            onChangeText={(text) => { this.textValueChanged(text, addInputItem) }}
-                            value={this.state.text}
-                        />
-                    </Input>
+					<TextInputWrapper 
+						json={this.payload}
+						handleFocus={this.handleFocus}
+						handleBlur={this.handleBlur}
+						textValueChanged={ (text) => 
+							{ this.textValueChanged(text, addInputItem) }}
+						value={this.state.text}    
+						isError={this.state.isError} 
+                    />
                 )}
             </InputContextConsumer>
         );
-    }
-
-    /**
-     * @description Return the input styles applicable based on the given payload
-     */
-    getComputedStyles = () => {
-        const { isMultiline } = this;
-
-        let inputComputedStyles = [styles.input, this.styleConfig.fontConfig];
-        isMultiline ?
-            inputComputedStyles.push(styles.multiLineHeight) :
-            inputComputedStyles.push(styles.singleLineHeight);
-        this.state.isError ?
-            inputComputedStyles.push(this.styleConfig.borderAttention) :
-            inputComputedStyles.push(styles.withBorderColor);
-
-        return inputComputedStyles;
     }
 
     /**
@@ -154,23 +101,6 @@ export class InputText extends React.Component {
     };
 
     /**
-     * @description Parse hostconfig specific to this element
-     */
-    parseHostConfig = () => {
-        this.id = this.payload.id;
-        this.type = this.payload.type;
-        this.isMultiline = this.payload.isMultiline;
-        this.maxLength = this.payload.maxLength;
-        this.placeholder = this.payload.placeholder;
-        let styleValue = Utils.parseHostConfigEnum(
-            Enums.InputTextStyle,
-            this.payload.style,
-            Enums.InputTextStyle.Text);
-        this.style = Utils.getEffectiveInputStyle(styleValue);
-        this.keyboardType = Utils.getKeyboardType(styleValue);
-    }
-
-    /**
      * @description handle textinput when in focus
      */
     handleFocus = () => {
@@ -186,25 +116,3 @@ export class InputText extends React.Component {
         this.validate();
     }
 }
-
-const styles = StyleSheet.create({
-    withBorderColor: {
-        borderColor: Constants.LightGreyColor,
-    },
-    multiLineHeight: {
-        height: 88,
-    },
-    singleLineHeight: {
-        height: 44,
-    },
-    input: {
-        width: Constants.FullWidth,
-        padding: 5,
-        marginTop: 15,
-        borderWidth: 1,
-        backgroundColor: Constants.WhiteColor,
-        borderRadius: 5,
-    },
-});
-
-
